requirements.txt
numpy==1.16.5
pandas==0.25.1
Pillow==9.0.0
scikit_learn==1.0.2
streamlit==1.4.0
